Update ClassPagesPrototype1 (1).html
Semifinal version! YAAY! However, this prototype does not yet import its data or format itself into a side-window. Nonetheless, this demonstrates basic functionality!!! Everyone, feel free to add to the CSS styles throughout if you have an idea of how to make it look better. In the touchup phase, we should use some SVG and images to make the page look professional. Anyway, this will work for the project increment! Let's definitely point out the limitations though so that we don't oversell this.